+++ exchangelib/autodiscover/cache.py
@@ -43,10 +43,11 @@ def shelve_open_with_failover(filename):
         shelve_handle = shelve.open(filename)
         # Try to actually use the shelve. Some implementations may allow opening the file but then throw
         # errors on access.
-        _ = shelve_handle['']
-    except KeyError:
-         # The entry doesn't exist. This is expected.
-        pass
+        try:
+            _ = shelve_handle['']
+        except KeyError:
+            # The entry doesn't exist. This is expected.
+            pass
     except Exception as e:
         for f in glob.glob(filename + '*'):
             log.warning('Deleting invalid cache file %s (%r)', f, e)